+++ src/scripts/models/news.js
